Move ModelResult from Microsoft.SemanticKernel.Orchestration to Microsoft.SemanticKernel.AI
This is the only type now in the Orchestration namespace. We should just remove that namespace entirely. We should subsequently decide whether it's worth keeping the "AI" namespace or removing that as well.
