+++ dotnet/src/Experimental/Orchestration.Flow.IntegrationTests/CollectEmailPlugin.cs
@@ -9,7 +9,6 @@ using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.AI;
 using Microsoft.SemanticKernel.AI.ChatCompletion;
 using Microsoft.SemanticKernel.Connectors.AI.OpenAI;
-using Microsoft.SemanticKernel.Orchestration;
 
 #pragma warning disable SKEXP0001
