+++ dotnet/src/Experimental/Orchestration.Flow.IntegrationTests/SendEmailPlugin.cs
@@ -3,7 +3,7 @@
 using System.ComponentModel;
 using System.Text.Json;
 using Microsoft.SemanticKernel;
-using Microsoft.SemanticKernel.Orchestration;
+using Microsoft.SemanticKernel.AI;
 
 namespace SemanticKernel.Experimental.Orchestration.Flow.IntegrationTests;
