+++ dotnet/src/Experimental/Orchestration.Flow/Abstractions/IFlowExecutor.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
 using System.Threading.Tasks;
-using Microsoft.SemanticKernel.Orchestration;
+using Microsoft.SemanticKernel.AI;
 
 namespace Microsoft.SemanticKernel.Experimental.Orchestration.Abstractions;
