+++ dotnet/src/Planners/Planners.Core.UnitTests/Action/ActionPlannerTests.cs
@@ -3,7 +3,7 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.SemanticKernel.AI;
 using Microsoft.SemanticKernel.AI.TextCompletion;
-using Microsoft.SemanticKernel.Orchestration;
+using Microsoft.SemanticKernel.AI;
 using Moq;
 using Xunit;
