+++ dotnet/src/Planners/Planners.Core.UnitTests/Planning/PlanTests.cs
@@ -2,8 +2,8 @@
 
 using System.Reflection;
 using Microsoft.Extensions.Logging;
+using Microsoft.SemanticKernel.AI;
 using Microsoft.SemanticKernel.Events;
-using Microsoft.SemanticKernel.Orchestration;
 using Microsoft.SemanticKernel.Planning;
 using Moq;
 using Xunit;
