+++ dotnet/src/Planners/Planners.Core.UnitTests/Sequential/SequentialPlannerTests.cs
@@ -3,7 +3,6 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.SemanticKernel.AI;
 using Microsoft.SemanticKernel.AI.TextCompletion;
-using Microsoft.SemanticKernel.Orchestration;
 using Moq;
 using Xunit;
